Use `failure` instead of `error-chain`
While not being as stable and battle-tested, it has more potential, uses
less magic than fully macro-based `error-chain` solution, is more
comfortable to work with and allows separation of errors originated from
different sources instead of assembling everything under the hood (this
function is fully covered by the `failure::Error` type).

Also this project is not a library one, so it doesn't need to pay much
attention to compatibility story.